fix: 변수 타입 오류 수정 Long -> integer

+++ services/bid-service/src/main/java/com/bankrupang/sanjijk/bid/presentation/dto/BidBroadcastDto.java
@@ -8,8 +8,8 @@ import lombok.Getter;
 public class BidBroadcastDto {
     private String type;
     private String auctionId;
-    private Long currentPrice;
+    private Integer currentPrice;
     private String previousHighestBidderId;
     private String highestBidderId;
-    private Long nextMinPrice;
+    private Integer nextMinPrice;
 }
